+++ fly_io/CHANGELOG.md
@@ -2,7 +2,15 @@
 
 <!-- towncrier release notes start -->
 
-## 1.0.0 / 2024-08-09
+## 1.0.1 / 2024-09-05
+
+***Fixed***:
+
+* Add more robust error handling. ([#18337](https://github.com/DataDog/integrations-core/pull/18337))
+* Support histogram metrics. ([#18338](https://github.com/DataDog/integrations-core/pull/18338))
+* Request `/v1/apps` for health check instead of the base URL. ([#18474](https://github.com/DataDog/integrations-core/pull/18474))
+
+## 1.0.0 / 2024-08-09 / Agent 7.57.0
 
 ***Added***: